LayoutTests/fast/scrolling/latching/iframe-latch-small-deltas-expected.txt


Tests that a second scroll with small x/y deltas uses latching from an earlier scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeTarget.contentWindow.pageYOffset is 0
PASS window.pageYOffset is 200
After scroll
PASS iframeTarget.contentWindow.pageYOffset is 500
PASS window.pageYOffset is 200
After wait
PASS iframeTarget.contentWindow.pageYOffset is 500
PASS window.pageYOffset is 200
PASS iframeTarget.contentWindow.pageYOffset is 480
PASS window.pageYOffset is 200
PASS successfullyParsed is true

TEST COMPLETE
